requirements.txt
biopython
pandas
numpy
scikit-learn
tensorflow
tqdm